requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.30.2
azure-storage-blob==12.22.0
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cx_Freeze==7.2.0
decorator==5.1.1
dmgbuild==1.6.2
dnspython==2.6.1
ds-store==1.3.1
ecdsa==0.19.0
fastapi==0.112.0
future==1.0.0
geocoder==1.38.1
h11==0.14.0
idna==3.7
isodate==0.6.1
jose==1.0.0
mac-alias==2.2.2
macholib==1.16.3
packaging==24.1
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pymongo==4.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
ratelim==0.1.6
requests==2.32.3
rsa==4.9
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wheel==0.43.0
